core upgrades to compile against boost 1.74 latest release

+++ src/addrman.cpp
@@ -391,8 +391,8 @@ CAddress CAddrMan::Select_(int nUnkBias)
     if (size() == 0)
         return CAddress();
 
-    double nCorTried = sqrt(nTried) * (100.0 - nUnkBias);
-    double nCorNew = sqrt(nNew) * nUnkBias;
+    double nCorTried = sqrt((double)nTried) * (100.0 - nUnkBias);
+    double nCorNew = sqrt((double)nNew) * nUnkBias;
     if ((nCorTried + nCorNew)*GetRandInt(1<<30)/(1<<30) < nCorTried)
     {
         // use a tried node
